requirements.txt
# Write down the modules you need to install and then
# run the cmd: ```pip install -r requirements.txt```
cryptography
python-dotenv